fix: metadata base options added to a root layout
