+++ app/[locale]/about/layout.tsx
@@ -6,7 +6,7 @@
 import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
-	title: 'About - Sundar Clinic',
+	title: 'About',
 	alternates: {
 		canonical: 'https://sundarclinic.com/about/',
 	},
