+++ app/[locale]/privacy-policy/layout.tsx
@@ -6,7 +6,7 @@
 import type { Metadata } from 'next';
 
 export const metadata: Metadata = {
-	title: 'Privacy Policy - Sundar Clinic',
+	title: 'Privacy Policy',
 	alternates: {
 		canonical: 'https://sundarclinic.com/privacy-policy/',
 	},
