+++ locales/hi.json
@@ -6,7 +6,11 @@
       "morning": "सुबह ‣ 9:30 बजे - 1:30 बजे",
       "evening": "शाम ‣ 4:00 बजे - 9:30 बजे"
     },
-    "address": "1195ए, नेहरू स्ट्रीट, चेन्नई, बेंगलुरु - चेन्नई हाईवे, पाप्पनछतिर म्, चेन्नई, तमिलनाडु - 600123।"
+    "address": "1195ए, नेहरू स्ट्रीट, चेन्नई, बेंगलुरु - चेन्नई हाईवे, पाप्पनछतिर म्, चेन्नई, तमिलनाडु - 600123।",
+    "meta": {
+      "description": "न केवल बेहतर स्वास्थ्य सेवा, बल्कि बैंगलोर-चेन्नई राजमार्ग पर स्थित पप्पनचतिराम में बेहतर स्वास्थ्य सेवा का अनुभव भी मिलता है। \nएक विश्वसनीय सामान्य चिकित्सक डॉ. एकता भारती के नेतृत्व में।",
+      "twitter": "बैंगलोर-चेन्नई राजमार्ग पर स्थित पप्पनचतिराम में सुंदर क्लिनिक में बेहतर स्वास्थ्य देखभाल और बेहतर स्वास्थ्य अनुभव की खोज करें। \nआपकी विश्वसनीय सामान्य चिकित्सक डॉ. एकता भारती के नेतृत्व में।"
+    }
   },
   "layouts": {
     "navbar": {
