+++ locales/ta.json
@@ -6,7 +6,11 @@
       "morning": "காலை ‣ 9:30 மணி - 1:30 மணி",
       "evening": "மாலை ‣ 4:00 மணி - 9:30 மணி"
     },
-    "address": "1195A, நேரு தெரு, சென்னை, பெங்களூரு - சென்னை நெடுஞ்சாலை, பாப்பன்சத்திரம், சென்னை, தமிழ்நாடு - 600123."
+    "address": "1195A, நேரு தெரு, சென்னை, பெங்களூரு - சென்னை நெடுஞ்சாலை, பாப்பன்சத்திரம், சென்னை, தமிழ்நாடு - 600123.",
+    "meta": {
+      "description": "பெங்களூர்-சென்னை நெடுஞ்சாலையில் அமைந்துள்ள பாப்பன்சத்திரத்தில் சிறந்த சுகாதாரம் மட்டுமல்ல, சிறந்த சுகாதார அனுபவமும் உள்ளது. \nநம்பகமான பொது மருத்துவர் டாக்டர் ஏக்தா பார்தி தலைமையில்.",
+      "twitter": "பெங்களூர்-சென்னை நெடுஞ்சாலையில் அமைந்துள்ள பாப்பன்சத்திரத்தில் உள்ள சுந்தர் கிளினிக்கில் சிறந்த சுகாதாரம் மற்றும் சிறந்த சுகாதார அனுபவத்தைக் கண்டறியவும். \nஉங்கள் நம்பகமான பொது மருத்துவர் டாக்டர் ஏக்தா பார்தி தலைமையில்."
+    }
   },
   "layouts": {
     "navbar": {
