+++ package.json
@@ -1,60 +1,63 @@
 {
-  "name": "sundar-clinic-main-website",
-  "version": "0.1.0",
-  "private": true,
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@hookform/resolvers": "^3.3.1",
-    "@mui/material": "^5.14.10",
-    "@portabletext/react": "^3.0.7",
-    "@radix-ui/react-accordion": "^1.1.2",
-    "@radix-ui/react-dialog": "^1.0.4",
-    "@radix-ui/react-label": "^2.0.2",
-    "@radix-ui/react-slot": "^1.0.2",
-    "@radix-ui/react-toast": "^1.1.4",
-    "@radix-ui/react-tooltip": "^1.0.6",
-    "@sanity/client": "^6.4.9",
-    "@sanity/image-url": "^1.0.0",
-    "@sanity/vision": "^3.0.0",
-    "@types/node": "20.5.9",
-    "@types/react": "18.2.21",
-    "@types/react-dom": "18.2.7",
-    "autoprefixer": "10.4.15",
-    "axios": "^1.5.0",
-    "class-variance-authority": "^0.7.0",
-    "clsx": "^2.0.0",
-    "eslint": "8.48.0",
-    "eslint-config-next": "13.4.19",
-    "lucide-react": "^0.274.0",
-    "next": "13.4.19",
-    "next-intl": "^3.20.0",
-    "next-sanity": "^5.0.0",
-    "nodemailer": "^6.9.5",
-    "postcss": "8.4.29",
-    "react": "18.2.0",
-    "react-dom": "18.2.0",
-    "react-hook-form": "^7.46.1",
-    "sanity": "^3.0.0",
-    "schema-dts": "^1.1.2",
-    "server-only": "^0.0.1",
-    "sharp": "^0.32.6",
-    "styled-components": "~5.2.0",
-    "suspend-react": "^0.1.3",
-    "tailwind-merge": "^1.14.0",
-    "tailwindcss": "3.3.3",
-    "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.2.2",
-    "zod": "^3.21.4"
-  },
-  "devDependencies": {
-    "@tailwindcss/typography": "^0.5.9",
-    "@types/nodemailer": "^6.4.10"
-  }
+	"name": "sundar-clinic-main-website",
+	"version": "0.1.0",
+	"private": true,
+	"scripts": {
+		"dev": "next dev",
+		"build": "next build",
+		"start": "next start",
+		"lint": "next lint",
+		"build:analyze": "cross-env ANALYZE=true pnpm build"
+	},
+	"dependencies": {
+		"@emotion/react": "^11.11.1",
+		"@emotion/styled": "^11.11.0",
+		"@hookform/resolvers": "^3.3.1",
+		"@mui/material": "^5.14.10",
+		"@next/bundle-analyzer": "^14.2.13",
+		"@portabletext/react": "^3.0.7",
+		"@radix-ui/react-accordion": "^1.1.2",
+		"@radix-ui/react-dialog": "^1.0.4",
+		"@radix-ui/react-label": "^2.0.2",
+		"@radix-ui/react-slot": "^1.0.2",
+		"@radix-ui/react-toast": "^1.1.4",
+		"@radix-ui/react-tooltip": "^1.0.6",
+		"@sanity/client": "^6.4.9",
+		"@sanity/image-url": "^1.0.0",
+		"@sanity/vision": "^3.0.0",
+		"@types/node": "20.5.9",
+		"@types/react": "18.2.21",
+		"@types/react-dom": "18.2.7",
+		"autoprefixer": "10.4.15",
+		"axios": "^1.5.0",
+		"class-variance-authority": "^0.7.0",
+		"clsx": "^2.0.0",
+		"eslint": "8.48.0",
+		"eslint-config-next": "13.4.19",
+		"lucide-react": "^0.274.0",
+		"next": "13.4.19",
+		"next-intl": "^3.20.0",
+		"next-sanity": "^5.0.0",
+		"nodemailer": "^6.9.5",
+		"postcss": "8.4.29",
+		"react": "18.2.0",
+		"react-dom": "18.2.0",
+		"react-hook-form": "^7.46.1",
+		"sanity": "^3.0.0",
+		"schema-dts": "^1.1.2",
+		"server-only": "^0.0.1",
+		"sharp": "^0.32.6",
+		"styled-components": "~5.2.0",
+		"suspend-react": "^0.1.3",
+		"tailwind-merge": "^1.14.0",
+		"tailwindcss": "3.3.3",
+		"tailwindcss-animate": "^1.0.7",
+		"typescript": "5.2.2",
+		"zod": "^3.21.4"
+	},
+	"devDependencies": {
+		"@tailwindcss/typography": "^0.5.9",
+		"@types/nodemailer": "^6.4.10",
+		"cross-env": "^7.0.3"
+	}
 }
